requirements.txt
# requirements.txt

# 核心依赖
requests==2.32.0
pytest==7.4.0


# 新增 WebSocket 和异步支持
websockets>=12.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0

# 数据处理
python-dateutil==2.8.2

# 日志和报告
allure-pytest==2.13.2
